initial logic implementation

+++ symtest.Common/Models/HttpRequestTemplate.cs
@@ -9,7 +9,7 @@ namespace symtest.Common.Models
         public HttpMethod Method { get; set; }
         
         public int Density { get; set; }
-        public int Duration { get; set; }
+        public double Duration { get; set; }
         public double Distribution { get; set; }
         
         public Dictionary<string, string> Headers { get; set; }
